fix: prdict 라우터로 분할
